refactor: separate productDetail from PricingPage
